+++ artemis-protocols/artemis-openwire-protocol/src/main/java/org/apache/activemq/artemis/core/protocol/openwire/amq/AMQConsumer.java
@@ -199,7 +199,7 @@ public class AMQConsumer {
 
             boolean topicChanged = !oldTopicName.equals(address);
 
-            if (selectorChanged || topicChanged) {
+            if ((selectorChanged || topicChanged) && !result.isConfigurationManaged()) {
                // Delete the old durable sub
                session.getCoreSession().deleteQueue(queueName);
